Refactor: divided one file into several pieces
