+++ run_glue.py
@@ -25,6 +25,11 @@ from dataclasses import dataclass, field
 from typing import Optional
 from accelerate import Accelerator
 
+from model_arguments import ModelArguments
+from data_training_arguments import DataTrainingArguments
+import constants
+
+import inspect
 import datasets
 import numpy as np
 from datasets import load_dataset, load_metric
@@ -60,203 +65,9 @@ check_min_version("4.17.0")
 
 require_version("datasets>=1.8.0", "To fix: pip install -r examples/pytorch/text-classification/requirements.txt")
 
-task_to_keys = {
-    "mnli": ("premise", "hypothesis"),
-    "mrpc": ("sentence1", "sentence2"),
-    "qnli": ("question", "sentence"),
-    "qqp": ("question1", "question2"),
-    "rte": ("sentence1", "sentence2"),
-    "sst2": ("sentence", None),
-}
-
-acc_tasks = ["mnli","qnli", "rte", "sst2"]
-f1_tasks = ["mrpc", "qqp"]
 
 logger = logging.getLogger(__name__)
 
-@dataclass
-class DataTrainingArguments:
-    """
-    Arguments pertaining to what data we are going to input our model for training and eval.
-
-    Using `HfArgumentParser` we can turn this class
-    into argparse arguments to be able to specify them on
-    the command line.
-    """
-
-    task_name: Optional[str] = field(
-        default=None,
-        metadata={"help": "The name of the task to train on: " + ", ".join(task_to_keys.keys())},
-    )
-    dataset_name: Optional[str] = field(
-        default=None, metadata={"help": "The name of the dataset to use (via the datasets library)."}
-    )
-    dataset_config_name: Optional[str] = field(
-        default=None, metadata={"help": "The configuration name of the dataset to use (via the datasets library)."}
-    )
-    max_seq_length: int = field(
-        default=128,
-        metadata={
-            "help": "The maximum total input sequence length after tokenization. Sequences longer "
-            "than this will be truncated, sequences shorter will be padded."
-        },
-    )
-    overwrite_cache: bool = field(
-        default=False, metadata={"help": "Overwrite the cached preprocessed datasets or not."}
-    )
-    pad_to_max_length: bool = field(
-        default=True,
-        metadata={
-            "help": "Whether to pad all samples to `max_seq_length`. "
-            "If False, will pad the samples dynamically when batching to the maximum length in the batch."
-        },
-    )
-    max_train_samples: Optional[int] = field(
-        default=None,
-        metadata={
-            "help": "For debugging purposes or quicker training, truncate the number of training examples to this "
-            "value if set."
-        },
-    )
-    max_eval_samples: Optional[int] = field(
-        default=None,
-        metadata={
-            "help": "For debugging purposes or quicker training, truncate the number of evaluation examples to this "
-            "value if set."
-        },
-    )
-    max_predict_samples: Optional[int] = field(
-        default=None,
-        metadata={
-            "help": "For debugging purposes or quicker training, truncate the number of prediction examples to this "
-            "value if set."
-        },
-    )
-    train_file: Optional[str] = field(
-        default=None, metadata={"help": "A csv or a json file containing the training data."}
-    )
-    validation_file: Optional[str] = field(
-        default=None, metadata={"help": "A csv or a json file containing the validation data."}
-    )
-    test_file: Optional[str] = field(default=None, metadata={"help": "A csv or a json file containing the test data."})
-
-    def __post_init__(self):
-        if self.task_name is not None:
-            self.task_name = self.task_name.lower()
-            if self.task_name not in task_to_keys.keys():
-                raise ValueError("Unknown task, you should pick one in " + ",".join(task_to_keys.keys()))
-        elif self.dataset_name is not None:
-            pass
-        elif self.train_file is None or self.validation_file is None:
-            raise ValueError("Need either a GLUE task, a training/validation file or a dataset name.")
-        else:
-            train_extension = self.train_file.split(".")[-1]
-            assert train_extension in ["csv", "json"], "`train_file` should be a csv or a json file."
-            validation_extension = self.validation_file.split(".")[-1]
-            assert (
-                validation_extension == train_extension
-            ), "`validation_file` should have the same extension (csv or json) as `train_file`."
-
-@dataclass
-class ModelArguments:
-    """
-    Arguments pertaining to which model/config/tokenizer we are going to fine-tune from.
-    """
-
-    model_name_or_path: str = field(
-        metadata={"help": "Path to pretrained model or model identifier from huggingface.co/models"}
-    )
-    config_name: Optional[str] = field(
-        default=None, metadata={"help": "Pretrained config name or path if not the same as model_name"}
-    )
-    tokenizer_name: Optional[str] = field(
-        default=None, metadata={"help": "Pretrained tokenizer name or path if not the same as model_name"}
-    )
-    cache_dir: Optional[str] = field(
-        default=None,
-        metadata={"help": "Where do you want to store the pretrained models downloaded from huggingface.co"},
-    )
-    use_fast_tokenizer: bool = field(
-        default=True,
-        metadata={"help": "Whether to use one of the fast tokenizer (backed by the tokenizers library) or not."},
-    )
-    model_revision: str = field(
-        default="main",
-        metadata={"help": "The specific model version to use (can be a branch name, tag name or commit id)."},
-    )
-    use_auth_token: bool = field(
-        default=False,
-        metadata={
-            "help": "Will use the token generated when running `transformers-cli login` (necessary to use this script "
-            "with private models)."
-        },
-    )
-
-    # kd setting
-    alpha_kd: float = field(
-        default=1.0,
-        metadata={
-            "help": "The weight of kd loss"
-        },
-    )
-    # mode: str = field(
-    #     default="kd",
-    #     metadata={"help": "The type of kd loss"},
-    # )
-    temperature: float = field(
-        default=1,
-        metadata={
-            "help": "The temperature."
-        },
-    )
-
-    # teacher model setting
-    teacher_model: str = field(
-        default=None,
-        metadata={
-            "help": "Path of teacher model."
-        },
-    )
-    train_teacher: bool = field(
-        default=False,
-        metadata={
-            "help": "Train teacher or not."
-        },
-    )
-    t_alpha_kd: float = field(
-        default=0.4,
-        metadata={
-            "help": "The weight of kd loss if train_teacher is True."
-        },
-    )
-    t_learning_rate: float = field(
-        default=3e-5,
-        metadata={
-            "help": "The learning rate of teacher."
-        },
-    )
-
-    # lgtm setting
-    use_lgtm: bool = field(
-        default=False,
-        metadata={
-            "help": "Use LGTM or not."
-        },
-    )
-    init_classifier_to_zero: bool = field(
-        default=False,
-        metadata={
-            "help": "Initialize the classifier of the teacher and student to zero."
-        },
-    )
-    num_layers: int = field(
-        default=6,
-        metadata={
-            "help": "The layer number of the student model."
-        },
-    )
-
-import inspect
 
 def _remove_unused_columns(model, dataset: "datasets.Dataset", description: Optional[str] = None):
     # if not self.args.remove_unused_columns:
@@ -444,7 +255,7 @@ def main():
 
     # Preprocessing the raw_datasets
     if data_args.task_name is not None:
-        sentence1_key, sentence2_key = task_to_keys[data_args.task_name]
+        sentence1_key, sentence2_key = constants.task_to_keys[data_args.task_name]
     else:
         # Again, we try to have some nice defaults but don't hesitate to tweak to your use case.
         non_label_column_names = [name for name in raw_datasets["train"].column_names if name != "label"]
@@ -756,9 +567,9 @@ def main():
                 for key, value in eval_metric.items():
                     logger.info(f" eval_{key}:{value} ")
            
-                if data_args.task_name in acc_tasks:
+                if data_args.task_name in constants.acc_tasks:
                     metric_key = "accuracy"
-                elif data_args.task_name in f1_tasks:
+                elif data_args.task_name in constants.f1_tasks:
                     metric_key = "f1"
                 
                 if eval_metric[metric_key] > best_metric:
